always load saved results when in uncompiled state
Previously, when a Procedure was in uncompiled state, it checksum
differed from the last known compile-state and therefore was marked for
re-parsing. As `FunctionList` relies on a compiled procedure, the
functionlist was left empty. This left CodeBrowser unusable when in an
uncompiled state.

This change skips checksum comparison when in uncompiled state and loads
the procedure from the save waves. As a consequence, the procedure's mark
is left in an invalid state.