fix inconsistency
since Function parsing is the first thing in parseProcedure(),
numEntries=0 and idx is basically starting to run at idx=0 in the for
loop. Therefore, function parsing worked as expected.

This is unwanted behaviour, as the function depends on being called first.
If not, not all list items from `funcList` will be read in `func =
StringFromList(idx, funcList)`. Depending on numEntries in declWave.

Changing this to a for loop that parses all functions no matter of the
current length of declWave.