add function comments
The comments above the function definition are displayed as help text.
This feature uses QT's helptext and requires Igor Pro 8.

The function marker description is added to the first column and the
comment to the second.

The  output of the tooltip is formatted with basic html.
The comment is enclosed in code tags to have equal line spacing.

support for Doxygen comments:
* @keyword is emphasized italic
* @param keyword is printed strong bold.