Unify position and resizing of panel elements
The codebrowser panel elements were not displayed nicely on windows 10.
The panel border of 5 was fixed and a new panel width of 300 was set.
The guide lines are adjusted so that the control elements above the
listbox get a fixed space defined by a guide line and the listbox below
remains adjustable.
All control elements are aligned to the right hand side. Since igor does
not support this directly, the controls were manually positioned.
New userdata needed to be written to ResizeControls