IgorBeforeQuitHook: Remove it to make it safer
We have a report of a user where the experiment was saved although it
should not have been.

This could not be reproduced here, but still the current approach is
inherently dangerous.

We now store and restore the experiment modified state surrounding the
panel closing and cleanup.

This does not make the Codebrowser panel closed in the saved experiment
nor does it cleanup its internal data. But we can live with these
drawbacks.

Reported-by: chozo at https://www.wavemetrics.com/comment/21323#comment-21323